PB-114: Moved ol coordinates extraction utils in api.feature
